services/python/python-service/greeter_server.py
from concurrent import futures
import logging

import grpc
import os
# These files below are genereated by protocol buffers compiler on Dockerfile.
import helloworld_pb2
import helloworld_pb2_grpc

class Greeter(helloworld_pb2_grpc.GreeterServicer):

    def SayHello(self, request, context):
        return helloworld_pb2.HelloReply(message='Hello, %s!' % request.name)


def serve():
    server = grpc.server(futures.ThreadPoolExecutor(max_workers=10))
    helloworld_pb2_grpc.add_GreeterServicer_to_server(Greeter(), server)
    port = os.getenv("PORT", "50051")
    server.add_insecure_port('[::]:'+port)
    server.start()
    server.wait_for_termination()


if __name__ == '__main__':
    logging.basicConfig()
    serve()